Project Structure: Move and rename Game classes to module
